Improved conditionally loading of responders
