+++ lib/response_manager/version.rb
@@ -1,3 +1,3 @@
 module ResponseManager
-  VERSION = "1.0.0"
+  VERSION = "2.0.0"
 end
